feat: Remove need for delays
Instead of relying on broacasted message to call the right method
on the HomePage view (which means we have to wait for this view to
appear), we simply set an attribute on the cozy service. When the
homepage appears (through OnAppearing), we can check this attribute.
We still broadcast a message, to force a check on this value in case
the OnAppearing callback is not called on resume.