fix: Persist `ClientId` and `RegistrationAccessToken` for logout
We want the Cozy's OAuth client to be deleted when the user logs out
from the app

With current implementation, the OAuth client can be deleted only if
the logout is done in the same runtime session than when the login has
been done. If the app is closed, then the `clientId` and
`registrationAccessToken` are lost and we won't be able to delete it
anymore

While `clientId` can be recovered from the persisted token, the
`registrationAccessToken` cannot be recovered. It is definitively lost

To prevent this, we want to persist this info

Related commit: e0ffa2fe6a41f020c4e17adf28ca45b203fcce76