fix: Make deep linking work for both scheme links and universal links
Scheme links are in the form of `cozypass://cozypass/action?params`. So
the URL's path is `/action`

Universal links are in the form of
`https://links.mycozy.cloud/pass/action?params`. So the URL's path is
`pass/action`

So we should remove the `/pass` value when existing to homogenize the
action interception