fix: Use universal links on LoginByEmail scenario
In #86 we reworked the login worflow in order to display the Cloudery
web pages instead of a native form. This would allow the user to login
by entering their email adresse. The received email would then be
opened in a browser which would then redirect to a `cozypass://` scheme

With this scenario, in some environments (i.e. Firefox mobile browser)
the user would be redirected to their Cozy login web form instead of
being redirect to the Cozy Pass mobile application

This is due to the manager's JS code that would fail opening the
`cozypass://` scheme and then redirects to the web login form

In #91 we implemented support to Universal Links, but we did not change
the Cloudery url to use them instead of the Scheme

By changing the scheme of an Universal Link, then the login by email
scenario would work in more environments like Firefox. This is because
the UL redirection is done at HTTP level instead of using a JS script
like it was done for Schemes. Also the final fallback is now the
`cozypass://` scheme instead of the web login form

This does not fix ALL scenario, in some rare case we can miss some
redirection. This is why we changed de Flagship app login to receive an
Universal Link directly in the email instead of a manager link
(cf cozy/cozy-flagship-app#834)
This may be done in the future, for now this implementation is not
prioritized

Related PR: #86
Related PR: #91
Related PR: cozy/cozy-flagship-app#834