fix: Do not crash when displaying collection header on iOS 18
We need now to build with iOS 18 SDK. But vault page with cipher list is crashing on iOS because a misuse of collection header in .NET. This issue has been fixed in .NET 8 which we can not use because we use Xamarin that is outdated.

So here we rework the way we display collection header and collection item by displaying only items. Some items have header styles but they are not official collection header anymore.

Related links
- https://github.com/dotnet/maui/issues/25139
- https://github.com/dotnet/maui/pull/25030/files