fix: Correct status bar color if default theme
When using default theme, we use "Cozy" theme or "Dark" theme depending if dark mode is enabled. "Cozy" theme is the only theme using light statusbar, that's why we force it manually. And we force it too when default theme is used. But when default theme is used and dark mode is set, we need to force dark statusbar.