fix: Send hash in OIDC login
During the SSO login, we still need to enter the master password even
if we can be considered as logged in (= having access tokens).

Before, we were never checking if the master password entered was
correct so the app was raising an crypto error
javax.crypto.BadPaddingException when trying to decipher the vault.
So here we send the hash password during the SSO login to be sure
that when we have our access tokens, the master password is also the
good one.

See https://github.com/cozy/cozy-stack/pull/4566